requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.13.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
babel=2.14.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
boruta=0.3=pypi_0
ca-certificates=2024.3.11=haa95532_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
cloudpickle=3.0.0=pypi_0
cmdstanpy=1.2.2=pypi_0
colorama=0.4.6=pypi_0
colorlog=6.8.2=pypi_0
comm=0.2.2=pypi_0
contourpy=1.1.1=pypi_0
cupy-cuda11x=12.3.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fastapi=0.110.0=pypi_0
fastjsonschema=2.19.1=pypi_0
fastrlock=0.8.2=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.50.0=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
future=1.0.0=pypi_0
google-auth=2.29.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.62.1=pypi_0
h11=0.14.0=pypi_0
holidays=0.46=pypi_0
httpcore=1.0.4=pypi_0
httpx=0.27.0=pypi_0
hyperopt=0.2.7=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
ipykernel=6.29.3=pypi_0
ipython=8.12.3=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
json5=0.9.24=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.1=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.4=pypi_0
jupyter-server=2.13.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.1.5=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.25.4=pypi_0
kiwisolver=1.4.5=pypi_0
libffi=3.4.4=hd77b12b_0
lightning=2.2.1=pypi_0
lightning-utilities=0.11.0=pypi_0
mako=1.3.2=pypi_0
markdown=3.6=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.3=pypi_0
nbformat=5.10.3=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.1=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.24.4=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.0.13=h2bbff1b_0
optuna=3.6.0=pypi_0
overrides=7.7.0=pypi_0
packaging=24.0=pypi_0
pandas=2.0.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pypi_0
patsy=0.5.6=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py38haa95532_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.2.0=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
prophet=1.1.5=pypi_0
protobuf=5.26.1=pypi_0
psutil=5.9.8=pypi_0
pure-eval=0.2.2=pypi_0
py4j=0.10.9.7=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.21=pypi_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.2=pypi_0
python=3.8.19=h1aa4202_0
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=2.0.7=pypi_0
pytorch-forecasting=1.0.0=pypi_0
pytorch-lightning=2.2.1=pypi_0
pytorch-optimizer=2.12.0=pypi_0
pytz=2024.1=pypi_0
pywin32=306=pypi_0
pywinpty=2.0.13=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
referencing=0.34.0=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.18.0=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.2=pypi_0
setuptools=68.2.2=py38haa95532_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.28=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack-data=0.6.3=pypi_0
stanio=0.5.0=pypi_0
starlette=0.36.3=pypi_0
statsmodels=0.14.1=pypi_0
sympy=1.12=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.4.0=pypi_0
tinycss2=1.2.1=pypi_0
tomli=2.0.1=pypi_0
torch=2.2.2+cu121=pypi_0
torchaudio=2.2.2+cu121=pypi_0
torchmetrics=1.3.2=pypi_0
torchvision=0.17.2+cu121=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.1=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.7.0=pypi_0
werkzeug=3.0.2=pypi_0
wheel=0.41.2=py38haa95532_0
xgboost=2.0.3=pypi_0
yarl=1.9.4=pypi_0
zipp=3.18.1=pypi_0
